Chore/version clap app to crate version (#51)
* Move spawn_command utility fn to run_cmd action

* Use clap crate_version macro to determine clap app version
